utility function to create events listener
